forward pass

+++ src/main/java/pv021/function/activation/ActivationFunction.java
@@ -1,6 +1,6 @@
 package pv021.function.activation;
 
 public abstract class ActivationFunction {
-    public abstract float apply(float potential);
-    public abstract float applyDifferentiated(float potential);
+    public abstract double apply(double potential);
+    public abstract double applyDifferentiated(double potential);
 }
